requirements.txt
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.9.2
arrow==1.2.0
astroid==2.8.0
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bitarray==1.2.2
boto3==1.18.56
botocore==1.21.56
buddy==0.0.3
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.1
cytoolz==0.11.0
decorator==5.1.0
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
fastdiff==0.3.0
hexbytes==0.2.2
idna==3.3
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
ipython==7.28.0
isort==5.9.3
jedi==0.18.0
jmespath==0.10.0
jsonschema==3.2.0
lark==0.11.3
lazy-object-proxy==1.6.0
lru-dict==1.1.7
matplotlib-inline==0.1.3
mccabe==0.6.1
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.8.0
packaging==21.0
parsec==3.9
parsimonious==0.8.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
pluggy==1.0.0
prompt-toolkit==3.0.20
protobuf==3.18.1
ptyprocess==0.7.0
py==1.10.0
py-solc-x==1.1.1
pycryptodome==3.11.0
Pygments==2.10.0
pylint==2.11.1
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.26.0
rlp==2.0.1
s3transfer==0.5.0
semantic-version==2.8.5
six==1.16.0
snapshottest==0.6.0
solidity-parser==0.0.7
tabulate==0.8.9
termcolor==1.1.0
toml==0.10.2
toolz==0.11.1
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
varint==1.0.2
wasmer==1.0.0
wasmer-compiler-cranelift==1.0.0
wcwidth==0.2.5
web3==5.24.0
websockets==9.1
wrapt==1.12.1
yapf==0.31.0
yarl==1.7.0
